requirements.txt
pytube 
openai-whisper 
ffmpeg
yt-dlp